+++ src/components/Modal.jsx
@@ -1,15 +1,19 @@
 import ReactDom from "react-dom";
 
-export default function Modal({ open, children, onClose }) {
-  if (!open) return null;
+export default function Modal({ state }) {
+  const [modal, setModal] = state;
+
+  // safeguard
+  if (modal === null) return null;
+
   return ReactDom.createPortal(
     <>
       <div className="overlay-modal">
         <div className="modal">
           <div className="close-button">
-            <button onClick={onClose}>X</button>
+            <button onClick={() => setModal(null)}>X</button>
           </div>
-          {children}
+          {modal}
         </div>
       </div>
     </>,
